fix(ci): split oversized quillins module and add size gate to pre-commit
PR CI unit-tests was red on a third gate that pre-commit did not mirror:
GATE-11 module-size budget. quill/ui/main_frame_quillins.py reached 641
lines (over the 600 default cap) when the wizard feature landed.

Per the budget policy (ratchet may only decrease, never raise for new
growth), split rather than re-baseline:
- Extract the _EditorHostServices adapter (~100 lines) to a new module
  quill/ui/main_frame_quillins_host.py; main_frame_quillins.py is now 539
  lines, under the cap. Behavior-preserving: the class is a standalone
  helper, not part of the frozen MainFrame surface.
- network_egress_audit.py: update the reviewed egress key to the new path
  (ui/main_frame_quillins_host.py::fetch) so GATE-9 still recognises it.
- dialog_inventory.json: re-register the moved show_choices surface.

Hardening:
- .pre-commit-config.yaml: add the GATE-11 module-size budget hook so this
  class of failure is caught locally before it reaches CI.

Verified: pre-commit (all 6 hooks), module-size budget, network egress
audit, banned patterns, and the quillins UI suite (16 passed) are green.
