+++ tests/unit/ui/fixtures/dialog_inventory.json
@@ -147,7 +147,7 @@
   "quill/ui/main_frame_quillins.py::QuillinsMenuMixin._quillin_consent::wx.MessageDialog": "native",
   "quill/ui/main_frame_quillins.py::QuillinsMenuMixin.open_quillins_manager.on_remove::wx.MessageDialog": "native",
   "quill/ui/main_frame_quillins.py::QuillinsMenuMixin.open_quillins_manager::wx.Dialog": "hardened_custom",
-  "quill/ui/main_frame_quillins.py::_EditorHostServices.show_choices::wx.SingleChoiceDialog": "native",
+  "quill/ui/main_frame_quillins_host.py::_EditorHostServices.show_choices::wx.SingleChoiceDialog": "native",
   "quill/ui/main_frame_selection.py::SelectionMarksMixin.quill_key_selection_actions::wx.SingleChoiceDialog": "native",
   "quill/ui/main_frame_sessions.py::SessionsMixin.open_session::wx.FileDialog": "native",
   "quill/ui/main_frame_sessions.py::SessionsMixin.save_session::wx.FileDialog": "native",
